delete for company and client contact done

+++ apps/clientmanagement/src/app/clients/add-client/company-contacts-form/company-contacts-form.component.ts
@@ -128,7 +128,7 @@ clientalreadyExist=false;
         this.clientalreadyExist=true;
 
         }
-
+this.IsEdit
       });
       if(!this.IsEdit){
 
@@ -235,6 +235,7 @@ clientalreadyExist=false;
     // }
     // this.addClientStateService.updateCompanyContacts(this.comapanyContacts);
     // this.addContactForm.reset();
+
   }
   edit(index:number){
     for(let count=0;count<this.listData.length;count++){
@@ -302,10 +303,11 @@ clientalreadyExist=false;
       nzOkType: 'primary',
       nzOkDanger: true,
       nzOnOk: () => {
-        console.log("deleteeeeeeeeeeeeeeeeee hereeeeeeeeeeeeee")
-
+      if(this.updateClientStateService.isEdit)
+      {
         if(typeof this.updateClientStateService.UpdateClientData.CompanyContacts[i].Guid!=='undefined')
         {
+
         this.companyService.DeleteCompany(this.updateClientStateService.UpdateClientData.CompanyContacts[i].Guid).subscribe(
           (res:any)=>{
             if(res.ResponseStatus==='Success')
@@ -324,11 +326,18 @@ clientalreadyExist=false;
         );
         }
         else{
-          console.log("deleteeeeeeeeeeeeeeeeee hereeeeeeeeeeeeee")
+
           this.removeItem(element,i);
           this.notification.success("Company Deleted Successfully","",{nzPlacement:'bottomRight'}
           );
         }
+      }
+      else{
+
+        this.removeItem(element,i);
+        this.notification.success("Company Deleted Successfully","",{nzPlacement:'bottomRight'}
+        );
+      }
 
       },
       nzCancelText: 'No',
